Use actual booleans (not strings) for boolean object values
The PDF supplied in Issue #4 contained a series of images in CCITT format,
which have a few boolean decode parameters which change the expected data.
The supplied PDF had two of these parameters set to "false", but as bools
were being stored as the strings "true" and "false", these were both being
evaluated to true by PHP, resulting in these decode params being set
incorrectly in the output, as TCPDI expects it to be a boolean.