Fix incorrect return of xref entry if 0 and object not found
If the xref table lists an object as having an offset of 0 and the strpos
check returns false (i.e. the object is in an objstream or simply not
there), findObjectOffset was incorrectly returning 0 instead of false.
Changing the check to === to ensure that both values are of the same type
avoids this issue and allows the calling function to handle this case
properly.  This was discovered in the sample PDF provided with issue #5