Don't attempt to extract streams which are inside other streams
Fixes issue #8.  findObjectOffsets() was attempting to extract an object
stream which was part of a compressed object stream (an embedded PDF),
which resulted in the extraction failing as it wasn't a valid FlateDecode
stream.  Streams which fall within other streams are now skipped in the
findObjectOffsets function, so it no longer adds them to the objoffsets
array or tries to extract them.