Skip spaces (0x20) preceding CR/LF at start of stream
As per the ReportViewer sample PDF provided in #8, some PDFs have an extra
space before the CR/LF after the "stream" start token, which throws off
position calculations.  Handled using a separate strspn to ensure that any
stream which happens to begin with 0x20 is not broken by this change.